:sparkles: support commonjs
Switch from node-minify to rollup, support amd & commonjs builds
